+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
 ### Breaking Changes
 
+- Requirements - `quantify-core` requires `qcodes>=0.35` now. (!427)
+
 ### Merged branches and closed issues
 
+- Documentation - Remove dependency on directory-tree (!431)
 - Documentation - Sphinx build now compatible with qcodes==0.36.0 (!416, counterpart of quantify-scheduler!552)
 - Analysis - Make all analysis classes available from `quantify_core.analysis` (!418)
+- Data - Improve `set_datadir` and test style and readability (!419)
 - Added AttributeError to the exceptions explictly handled in loading of datasets for the remote plotmon (#352, !442)
 - Added exception handling for SI_utilities.value_precision (#350, !442)
 
@@ -383,4 +387,3 @@ Please see merge request !101 for a python snippet that will make all previous d
 - \#  denotes a closed issue.
 - ! denotes a merge request.
 
----